bot: Added specific commands
The telegram bot has the ability to respond to the implemented
commands. This allows sending the currently implemented commands
to be available on demand.

The code has been split in different files to fit commands in
their specific code file.

Copyright and license data have also been added to different files.