io: Fix function execution
The function execution given different summary types is crashing.
This is because the function is being called with more parameters
than supported by the function itself. Exactly the `summary_type`
is passed twice, one explicitly and the other by the expansion
of the `SUMMARY_DATA` dictionary due to last minute change.

This has been fixed by removing the explicit `summary_type`
passing.