project: Improve code to support future enhancements
The way that the bot handles the information has been changed. Due
to this, the source of the data has also be changed, and also the
code that imports and parses it.

This will allow custom queries of the data that could extend the
commands with particular arguments.

Message creation functions has also been removed from the `settings`
module.

The configuration file has also been changed to support properly
these changes.